改善: 出力ページのUIを完全にリニューアル
- 高さ制約問題の根本解決: CSS Grid/Flexboxによる動的レイアウト
- タブ化による出力タイプ別UI: 採点済み答案PDF/採点データExcel/個人成績表PDF
- 生徒選択UIの改善: コンパクトな3行レイアウト、状態フィルタのアイコン化
- 07-score-at-onceパターン適用: PageHeader内部管理によるlayout.tsx制約回避
- 固定値/calc()を一切使用しない純粋なflex/gridソリューション
